refactor: moved handler function files to appropriate folders
